requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
coverage==6.3.2
Django==4.0.3
django-bootstrap-v5==1.0.11
django-filter==21.1
idna==3.3
pyTelegramBotAPI==4.4.0
requests==2.27.1
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.8
